Add Labels to TaxonomiesService and Product view
Labels were partially implemented but still needed to be incorporated
into the API Models and Network Services. In doing so, it was discovered
that the Taxonomies weren't being downloaded to Realm on first launch of
the app as intended. After fixing that and adding the Label taxonomy to
Models, they were successfully retrieved and stored. Then, to allow
Labels to be displayed by ProductDetailViewController, they were
included in the OFFJson mapping. The downloaded taxonomies were then
used to translate Labels and code was added to fix translation of
Categories. Lastly, the functionality to link link Labels to their OFF
urls was added.

Closes #719