Create UI for ProductAttributes in Detail View
The product summary Overview needs to display rows for a product's
Attributes, whose data are fetched from the newly integrated
ProductAttribute API endpoint. This required a subclass of UIView and a
new class of TableView Cells to hold them. When the user taps an
Attribute row, a pop-up info card will be displayed showing a longer
description. Swiping left will show more related information on the
pop-up info card.

Resolves: #707