Attempt to show floatingPanel from closure
This snapshot is not workable, I just want to preserve an idea which
probably won't work, which is passing a closure to
AttributeTableViewCell to launch the floatingPanel. However, I made
FormTableViewController the owner of the floating panel and that is not
allowed. So now I am switching to a new strategy to make a delegate for
AttributeTableViewCell and designate the
ProductDetailTableViewController as the delegate and as the floating
panel host as I originally intended.