NOT READY FOR PROD. Wrap text around the image
Add the frame of the image to the exclusion path of the text view.
Create func getIconAspectConstraint() to get the dynamic aspect ratio of
the image in code so we can dynamically adjust the imageView height and
width. Place image related code in completion handle of KingFisher
setImage(with: resource) to ensure the adjustments are attempted after
the image is retrieved.

Resolves: #707