[Up for adoption] Bumping xcode to make the build work again (#904)
* Create UI for ProductAttributes in Detail View

The product summary Overview needs to display rows for a product's
Attributes, whose data are fetched from the newly integrated
ProductAttribute API endpoint. This required a subclass of UIView and a
new class of TableView Cells to hold them. When the user taps an
Attribute row, a pop-up info card will be displayed showing a longer
description. Swiping left will show more related information on the
pop-up info card.

Resolves: #707

* Display AttributeCells and increase attributed string font size

* Attempt to show floatingPanel from closure

This snapshot is not workable, I just want to preserve an idea which
probably won't work, which is passing a closure to
AttributeTableViewCell to launch the floatingPanel. However, I made
FormTableViewController the owner of the floating panel and that is not
allowed. So now I am switching to a new strategy to make a delegate for
AttributeTableViewCell and designate the
ProductDetailTableViewController as the delegate and as the floating
panel host as I originally intended.

* Create custom VC for ProdAttr. and use delegation

Created delegate protocol for AttributeTableViewCell to call back to
ProductDetailViewController so it can launch the FloatingPanelVC
Extend ProductDetailVC so it can handle FloatingPanels similar to
ScannerViewController. Create ProductAttributeViewController.swift and
.storyboard to be the FloatingPanels VC and view to hold AttributeView

Resolves: #707

* Display floating panel

* Dynamically update insets of floatingpanel

* Dynamically update FloatingPanel view insets

* Dismiss FloatingPanel from SummaryFormTableVC

* NOT READY FOR PROD. Wrap text around the image

Add the frame of the image to the exclusion path of the text view.
Create func getIconAspectConstraint() to get the dynamic aspect ratio of
the image in code so we can dynamically adjust the imageView height and
width. Place image related code in completion handle of KingFisher
setImage(with: resource) to ensure the adjustments are attempted after
the image is retrieved.

Resolves: #707

* Text wraps image but only 1st time

* Anchor icon in AttributeView to top corner so text wraps

* Switch ImageView to UIWebView, add code to scale SVGs, regex search in SVG's HTML to get dimensions

* Fix SVG icon not centered

* Add activity spinner to show loading Attr. Icon

* Fix AttributeView heighth on FloatingPanelVC

* Cleanup minor SwiftLint warnings

* Move activitySpinner back to top for visibility

* Fix mem leak: IngrdAnlysCell retained SummaryForm

* Update ios-release.yml

* Update ios-release.yml

* Update Gemfile.lock

* rename

d

* Update Gemfile.lock

* Update ios-release.yml

* Update Fastfile

* Create carthage-build.sh

* Update ios-release.yml

* Update Mintfile

* Update setup.sh

* Update setup.sh

* Update Fastfile

* Update setup.sh

* Update Fastfile

* Upgrade to Xcode 12 (#906)

* Update dependencies

* Update project.yml to use XCFrameworks

* Extract scripts/carthage.sh

* Make scripts/setup.sh work when $CI is not set

Co-authored-by: Alex Beaty <romeobeaty@gmail.com>
Co-authored-by: Slava Shklyaev <slava@slava.sh>